refactor(api): réorganise les middlewares
